+++ packages/three/src/index.ts
@@ -3,3 +3,6 @@ export * from "./GraphAnalyticsController";
 export * from "./GraphQLDataSource";
 export * from "./GraphPresentationController";
 export * from "./GraphMiniMap";
+export * from "./CanvasGraphRenderer";
+export * from "./GraphPluginRegistry";
+export * from "./MapLibreGraphLayer";